Remove legacy analytics module file
