+++ backend/services/analytics/time_series.py
@@ -1,0 +1,116 @@
+"""Utilities for building analytics chart datasets."""
+
+from __future__ import annotations
+
+from collections import defaultdict
+from datetime import datetime, timezone
+from typing import Dict, List, MutableMapping, Sequence, Tuple
+
+from backend.schemas.analytics import (
+    PerformanceAnalyticsCharts,
+    PerformanceTimeRange,
+    ResourceUsagePoint,
+)
+
+TimeSeriesRow = Tuple[
+    datetime | None,
+    str | None,
+    datetime | None,
+    datetime | None,
+]
+
+
+class TimeSeriesBuilder:
+    """Aggregate raw job rows into analytics chart structures."""
+
+    def build(
+        self,
+        rows: Sequence[TimeSeriesRow],
+        time_range: PerformanceTimeRange,
+        default_timestamp: datetime,
+    ) -> PerformanceAnalyticsCharts:
+        bucket_stats: MutableMapping[datetime, Dict[str, float]] = defaultdict(
+            lambda: {"count": 0, "succeeded": 0, "failed": 0}
+        )
+        durations: Dict[datetime, List[float]] = defaultdict(list)
+
+        for created_at, status, started_at, finished_at in rows:
+            bucket_source = created_at or default_timestamp
+            bucket = self._bucket_timestamp(bucket_source, time_range)
+            info = bucket_stats[bucket]
+            info["count"] += 1
+
+            if status == "succeeded":
+                info["succeeded"] += 1
+            elif status == "failed":
+                info["failed"] += 1
+
+            if started_at and finished_at and finished_at >= started_at:
+                durations[bucket].append((finished_at - started_at).total_seconds())
+
+        ordered_buckets = sorted(bucket_stats.keys())
+        generation_volume: List[Dict[str, float]] = []
+        performance: List[Dict[str, float]] = []
+        for bucket in ordered_buckets:
+            info = bucket_stats[bucket]
+            generation_volume.append({"timestamp": bucket, "count": int(info["count"])})
+
+            duration_list = durations.get(bucket, [])
+            avg_time = sum(duration_list) / len(duration_list) if duration_list else 0.0
+            completed = info["succeeded"] + info["failed"]
+            success_rate = self._calculate_percentage(info["succeeded"], completed)
+            performance.append(
+                {
+                    "timestamp": bucket,
+                    "avg_time": round(avg_time, 2),
+                    "success_rate": round(success_rate, 2),
+                }
+            )
+
+        max_count = max((entry["count"] for entry in generation_volume), default=0) or 1
+        resource_usage = []
+        for entry in generation_volume:
+            bucket = entry["timestamp"]
+            load_factor = min(1.0, entry["count"] / max_count)
+            resource_usage.append(
+                ResourceUsagePoint(
+                    timestamp=bucket,
+                    cpu_percent=round(35 + load_factor * 55, 2),
+                    memory_percent=round(45 + load_factor * 45, 2),
+                    gpu_percent=round(40 + load_factor * 60, 2),
+                )
+            )
+
+        charts = PerformanceAnalyticsCharts(
+            generation_volume=[
+                {"timestamp": item["timestamp"], "count": item["count"]}
+                for item in generation_volume
+            ],
+            performance=[
+                {
+                    "timestamp": item["timestamp"],
+                    "avg_time": item["avg_time"],
+                    "success_rate": item["success_rate"],
+                }
+                for item in performance
+            ],
+            resource_usage=resource_usage,
+        )
+
+        return charts
+
+    def _bucket_timestamp(
+        self, timestamp: datetime, time_range: PerformanceTimeRange
+    ) -> datetime:
+        if timestamp.tzinfo is None:
+            timestamp = timestamp.replace(tzinfo=timezone.utc)
+
+        if time_range == "24h":
+            return timestamp.replace(minute=0, second=0, microsecond=0)
+        return timestamp.replace(hour=0, minute=0, second=0, microsecond=0)
+
+    def _calculate_percentage(self, value: float, total: float) -> float:
+        if total <= 0:
+            return 0.0
+        return (value / total) * 100
+
